django/requirements.txt
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-jazzmin==2.4.8
django-redis==5.0.0
django-social-share==2.1.1
djangorestframework==3.12.2
mock==4.0.3
model-bakery==1.2.1
pillow==8.0.1
pip-chill==1.0.1
pre-commit==2.15.0
psycopg2-binary==2.9.2
python-dotenv==0.19.0
